Moved config out of src
